fix - sitemap generator
